(#180) Move commands descriptions to separate directory
