+++ src/commands/generator.cpp
@@ -1,9 +1,9 @@
+#include <iostream>
 #include <string>
 #include <fstream>
 #include <thread>
 #include <experimental/filesystem>
 #include "shell.h"
-#include "task.h"
 #include "utils.h"
 
 namespace comproenv {
